Moved ToastNotification and SearchBox
