+++ bookkeeping-app/src/components/contexts/ToastCtx.jsx
@@ -1,6 +1,6 @@
 import { createContext, useContext, useState, useCallback, useEffect } from "react";
 
-import ToastNotification from "../elements/misc/ToastNotification";
+import ToastNotification from "../elements/utilities/ToastNotification";
 
 const ToastCtx = createContext(null);
