+++ bookkeeping-app/src/pages/EntitiesPage.jsx
@@ -8,7 +8,7 @@ import penIcon from "../assets/pen-icon.svg";
 import TransactionItem from "../components/elements/items/TransactionItem";
 import AddEntityModal from "../components/elements/modals/AddEntityModal";
 import ConfirmationModal from "../components/elements/modals/ConfirmationModal";
-import SearchBox from "../components/elements/misc/SearchBox";
+import SearchBox from "../components/elements/utilities/SearchBox";
 import NoResultsDisplay from "../components/elements/utilities/NoResultsDisplay";
 import Input from "../components/elements/utilities/Input";
 import Button from "../components/elements/utilities/Button";
