+++ bookkeeping-app/src/pages/PropertiesPage.jsx
@@ -6,7 +6,7 @@ import PropertiesCtx from "../components/contexts/PropertiesCtx";
 import ConfirmationModal from "../components/elements/modals/ConfirmationModal";
 import penIcon from "../assets/pen-icon.svg";
 import AddPropertyModal from "../components/elements/modals/AddPropertyModal";
-import SearchBox from "../components/elements/misc/SearchBox";
+import SearchBox from "../components/elements/utilities/SearchBox";
 import RentInformation from "../components/elements/misc/RentInformation";
 import Input from "../components/elements/utilities/Input";
 import PropertyTypeDropdown from "../components/elements/dropdowns/PropertyTypeDropdown";
